+++ lib/dargres.dart
@@ -6,3 +6,5 @@ export 'src/query.dart';
 export 'src/ssl_context.dart';
 export 'src/transaction_context.dart';
 export 'src/exceptions.dart';
+export 'src/to_statement.dart';
+